Refactor type hints and imports for improved clarity and consistency across multiple files
